Evaluate and plot multiple impulse perturbation amplitudes in notebook 1-2b
- Vmap the evaluation of responses to multiple amplitudes of impulse perturbation in nb 1-2b.
- Estimate the memory that will be taken by all the impulse response states, before evaluating
  them.
- Adjust existing figures to map over impulse responses amplitudes.
- Rename a couple of minor things.
- Use tqdm to show progress on figure generation/saving in nb 1-2b.
- Refactor the ugly tree flatten with paths into a function, for looping over trees of figures.
- Start using custom-named tuples for easier rearrangement/transposition of PyTrees

The structure of the PyTrees in nb 1-2b is now kind of unwieldy. Perhaps I should have
relied more on vmap and less on `tree_map` to begin with, at least for tree levels
with identically sized arrays (i.e. most or all of them). Then I would have been able
to (say) use `jnp.moveaxis` rather than doing complicated things to shift the levels
of the PyTree around. Well, I'm not planning on changing it now. Let it be a lesson
for future analyses I perform.